Layout changes: * move utils to configmanager.utils module * move config item related stuff to configmanager.items * move config manager related stuff to configmanager.managers
